pkg/api/core: move pkg/api/endpoints
