+++ pkg/apis/core/helper/endpoints/BUILD
@@ -9,7 +9,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["util.go"],
-    importpath = "k8s.io/kubernetes/pkg/api/endpoints",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/endpoints",
     deps = [
         "//pkg/apis/core:go_default_library",
         "//pkg/util/hash:go_default_library",
@@ -21,7 +21,7 @@ go_test(
     name = "go_default_test",
     srcs = ["util_test.go"],
     embed = [":go_default_library"],
-    importpath = "k8s.io/kubernetes/pkg/api/endpoints",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/endpoints",
     deps = [
         "//pkg/apis/core:go_default_library",
         "//vendor/github.com/davecgh/go-spew/spew:go_default_library",
